# msclaw/docs/openapi.yaml
openapi: 3.1.0
info:
  title: MSClaw Control API
  description: |
    Control plane for MSClaw - Microsoft Security Operations Agent Runtime.
    Handles workflow submission, approval management, and audit trail queries.
  version: 0.1.0
  license:
    name: MIT

servers:
  - url: http://localhost:8100
    description: Local development

paths:
  /health:
    get:
      tags: [health]
      summary: Health check
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: control-api

  /ready:
    get:
      tags: [health]
      summary: Readiness check
      operationId: readinessCheck
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready

  /api/v1/workflows:
    post:
      tags: [workflows]
      summary: Submit a workflow for execution
      operationId: submitWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowSubmitRequest'
      responses:
        '200':
          description: Workflow submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSubmitResponse'
    get:
      tags: [workflows]
      summary: List workflow runs
      operationId: listWorkflows
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, awaiting_approval, completed, failed, cancelled]
      responses:
        '200':
          description: List of workflow runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowListItem'

  /api/v1/workflows/{runId}:
    get:
      tags: [workflows]
      summary: Get workflow run status
      operationId: getWorkflowStatus
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workflow run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStatusResponse'
        '404':
          description: Run not found

  /api/v1/workflows/{runId}/cancel:
    post:
      tags: [workflows]
      summary: Cancel a running workflow
      operationId: cancelWorkflow
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workflow cancelled
        '400':
          description: Cannot cancel in current status
        '404':
          description: Run not found

  /api/v1/approvals:
    get:
      tags: [approvals]
      summary: List approval requests
      operationId: listApprovals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            default: pending
            enum: [pending, approved, denied, expired]
      responses:
        '200':
          description: List of approval requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApprovalResponse'

  /api/v1/approvals/{approvalId}/decide:
    post:
      tags: [approvals]
      summary: Approve or deny a request
      operationId: decideApproval
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
        '400':
          description: Already decided or invalid decision
        '404':
          description: Approval not found

  /api/v1/audit:
    get:
      tags: [audit]
      summary: Query audit entries
      operationId: listAuditEntries
      parameters:
        - name: correlation_id
          in: query
          schema:
            type: string
        - name: workflow_run_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Audit entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEntryResponse'

  /api/v1/audit/chain/{correlationId}:
    get:
      tags: [audit]
      summary: Get full audit chain for a correlation ID
      operationId: getAuditChain
      parameters:
        - name: correlationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audit chain
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEntryResponse'

components:
  schemas:
    WorkflowSubmitRequest:
      type: object
      required: [workflow_id, inputs]
      properties:
        workflow_id:
          type: string
          example: contain-device-from-incident
        inputs:
          type: object
          example: { "incidentId": "INC-42" }
        idempotency_key:
          type: string
          format: uuid
        operator_id:
          type: string
          example: operator@contoso.com

    WorkflowSubmitResponse:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
        correlation_id:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string

    WorkflowStatusResponse:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
        workflow_id:
          type: string
        status:
          type: string
          enum: [pending, running, awaiting_approval, completed, failed, cancelled]
        steps_completed:
          type: array
          items:
            type: integer
        step_results:
          type: object
        artifacts:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error:
          type: string
          nullable: true

    WorkflowListItem:
      type: object
      properties:
        run_id:
          type: string
          format: uuid
        workflow_id:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        operator_id:
          type: string

    ApprovalDecisionRequest:
      type: object
      required: [decision, decided_by]
      properties:
        decision:
          type: string
          enum: [approved, denied]
        decided_by:
          type: string
          example: senior-op@contoso.com
        note:
          type: string
          nullable: true

    ApprovalResponse:
      type: object
      properties:
        approval_id:
          type: string
          format: uuid
        workflow_run_id:
          type: string
          format: uuid
        step_index:
          type: integer
        plugin:
          type: string
        action:
          type: string
        reason:
          type: string
        parameters:
          type: object
        status:
          type: string
          enum: [pending, approved, denied, expired]
        requested_at:
          type: string
          format: date-time
        decided_at:
          type: string
          format: date-time
          nullable: true
        decided_by:
          type: string
          nullable: true
        decision_note:
          type: string
          nullable: true

    AuditEntryResponse:
      type: object
      properties:
        entry_id:
          type: string
          format: uuid
        correlation_id:
          type: string
          format: uuid
        workflow_run_id:
          type: string
          format: uuid
          nullable: true
        action:
          type: string
        actor:
          type: string
        target:
          type: string
          nullable: true
        plugin:
          type: string
          nullable: true
        tool_action:
          type: string
          nullable: true
        inputs:
          type: object
          nullable: true
        outputs:
          type: object
          nullable: true
        policy_decision:
          type: string
          nullable: true
        model_used:
          type: string
          nullable: true
        approval_chain:
          type: array
          items:
            type: string
          nullable: true
        error:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
        entry_hash:
          type: string
          nullable: true